test: do not exports test utils
