+++ test/utils.js
@@ -1,30 +1,31 @@
 'use strict';
 
-const path = require('path');
 const Coffee = require('coffee').Coffee;
 
+// https://github.com/popomore/coffee
 class CliCoffee extends Coffee {
   constructor(...args) {
     super(...args);
-    this.promptActions = [];
-    this.promptMockScript = path.join(__dirname, './mock-script.js');
+    this.prompts = [];
   }
 
-  mockPrompt(input) {
-    this.promptActions.push(input);
+  prompt(input) {
+    this.prompts.push(input);
     return this;
   }
 
-  get _hookScripts() {
-    const str = super._hookScripts;
-    if (this.promptActions.length === 0) return str;
-
-    const mockScript = `require('${this.promptMockScript}')(${JSON.stringify(this.promptActions)})`;
-    return `${str || ''}\n\n${mockScript}`;
+  _run() {
+    const res = super._run();
+    this.proc.on('message', ({ type }) => {
+      if (type !== 'prompt') return;
+      const key = this.prompts.shift();
+      this.proc.stdin.write(key + '\n');
+    });
+    return res;
   }
 
   restore() {
-    this.promptActions = [];
+    this.prompts = [];
     return super.restore();
   }
 }
